+++ lib/libzpool/Makefile.am
@@ -58,6 +58,7 @@ KERNEL_C = \
 	dmu_diff.c \
 	dmu_object.c \
 	dmu_objset.c \
+	dmu_recv.c \
 	dmu_send.c \
 	dmu_traverse.c \
 	dmu_tx.c \
