+++ module/zfs/Makefile.in
@@ -34,6 +34,7 @@ $(MODULE)-objs += dmu.o
 $(MODULE)-objs += dmu_diff.o
 $(MODULE)-objs += dmu_object.o
 $(MODULE)-objs += dmu_objset.o
+$(MODULE)-objs += dmu_recv.o
 $(MODULE)-objs += dmu_send.o
 $(MODULE)-objs += dmu_traverse.o
 $(MODULE)-objs += dmu_tx.o
